feat: nestCell convert a 2D cell array into a column of nested cells
This function essentially does the opposite of the flattedCell function. nestCell takes an NxM cell array and merges all the columns into an Nx1 column cell array, where each of the entries are nested cells with M entries. The user can optionally specify the removal of empty cells, where empty entries will not be included in the nested cells.