feat: mapMetsToAltModel improved flexibility of model fields to use
The user can now specify which model fields are to be used for mapping metabolite IDs, and the "metNames" field is optional (i.e., it will not be used unless specified as a field in the input). Also, the input option "prioritize" (previously "allmatches") allows the user to specify whether the provided model fields should be prioritized when mapping metabolites. This means that when a metabolite is mapped using one of the fields provided as input, it will no longer be mapped by subsequent fields. In this manner, the order in which the model fields are provided in the "metFields" input can impact the results. If the prioritization input is FALSE, then the function will return all metabolite matches across all the fields listed in metFields, regardless of order.