feat: addAltsToModelField new function, was originally subfunction
The addAltsToModelField function was originally a sub-function within the mapModelMets function, but given that it can be used for other purposes, it has been separated as an independent function. It compares and appends a new set of field values to an existing model field. If an existing field entry is empty, it will be overwritten by the new entry. If the existing field entry and new entry conflict, both will be saved by adding new columns to the model field.