feat: new inputs and improved user control
Users can now input their own gene ID conversion key. Furthermore, the user can explicitly specify the type of geneIDs in the input, if it is suspected that the function will not be able to properly detect the gene ID type. Finally, the input "model" was changed to simply "grRules", so that the user would not need to input an entire model structure, but just the grRules that are to be translated.