feat: combineModelGPRsScript new master script to update model GPRs
Master script for combining and integrating genes and grRules from iHsa, and Recon3D into the merged model. The script makes use of the "integrateGeneRules.m" and "translateGeneRules.m" functions, among others. 

The script also adds additional explicit protein-reaction relationships to the model, through the addition of three new fields: "proteins", "prRules", and "rxnProtMat", which are analogous to the "genes", "grRules", and "rxnGeneMat" fields. NOTE: the entries in the protein-related fields are NOT aligned with those in the gene-related fields, so they must not be treated as such.