feat: humanGEM updated genes, grRules, rxnGeneMat, and new protein info
The humanGEM (v0.2.0) was modified using the "combineModelGPRsScript" script, which updated the grRules with a merged combination of grRules from HMR2, iHsa, and Recon3D. The "genes" and "rxnGeneMat" fields were updated accordingly.

In addition, the script added three new protein-related fields to the model: "proteins", "prRules", and "rxnProtMat", which are analogous to the "genes", "grRules", and "rxnGeneMat" fields, respectively. These protein-related fields utilize UniProt IDs instead of Ensembl IDs.

NOTE: The protein-related fields DO NOT correspond (i.e., they are not "aligned") with the gene-related fields, so they should not be treated as such.