model: humanGEM added boundary metabolites to all relevant rxns
Boundary metabolites (those in the "x" compartment) were incorporated in the latest version of humanGEM using the "addBoundaryMets" function. This function found all reactions in the model that involved only a single, extracellular metabolite. A corresponding boundary metabolite was added to each of these reactions, to make it consistent with the RAVEN formatting of exchange rxns. If the model did not have a boundary version of the extracellular metabolite, it was created, and added to the model (using the "addMets" function).