feat: curated_CORUM_grRules_20180924 new grRules with added complex info
The grRules from HumanGEM were automatically processed using the "addComplexesToGeneRules" script, to add CORUM complex information. This process was restricted to only those grRules in HumanGEM that did NOT contain any "AND" expressions. 

The rules that were changed by integration of the CORUM complex information were subsequently manually curated to evaluate which of the newly added complexes were consistent with literature/databases.

The results of this manual curation were summarized in this txt file (curated_CORUM_grRules_20180924.txt), which contains the rxn ID, the original grRule, and the newly curated grRule for each rule that changed.