feat: fatty_acid_export_rxns.txt list of new transport rxns to add
Since the reactions involving degradation of two of the fatty acid pools were constrained, this resulted in an inability of the model to consume these pools. Therefore, reactions allowing export of each of the fatty acid pools will be added, and these are stored in a txt file for easier tracking/implementation.