feat: identifyMassVariableRxns new function to find mass-inconsistencies
Note: this function originated from a previous version of the "constrainVariableMassReactions.m" script. 

This function searches a model for sets of reactions that involve the same set of metabolites except for one, which indicates that one or more of the reactions in the set is mass-imbalanced. The metabolites that differ among the reactions in a set will be compared in their metFormulas; if they are different, the set will be classified as "mass variable reactions", otherwise the set will be classified as "mass consistent reactions".