refactor: curateBiomassReactions change location of biomass production
Biomass reactions were originally planned to be formulated such that they produced the "biomass" metabolite in the boundary compartment (x). However, since this sometimes confuses functions into thinking that the biomass reactions are exchange reactions, this was altered so that the biomass reactions now produce biomass in the cytoplasmic compartment. Two transport reactions were added to enable export of the biomass metabolite out of the system (to the boundary).