feat: metabolicTasks_LeakCheck.xls created task list for checking leaks
The Excel document can be used by RAVEN's checkTasks or parseTaskList functions to evaluate each of the metabolic tasks defined in the document for a given model. metabolicTasks_LeakCheck.xls contains 17 metabolic tasks that test for things such as generation of carbon and oxygen from nothing, phosphorylation of ADP (to ATP) with no inputs, and realistic ATP generation per glucose consumed.

This task list can therefore replace the existing "checkEnergyMassLeaks" function, as it is a more exhaustive and straightforward evaluation.