refactor: getINITModel2 incorporate temporary computation shortcut
This modification allows one to skip the (unnecessary) re-running of the simplifyModel and checkTasks functions prior to the tINIT algorithm when running this for several models. These values are instead calculated beforehand, and fed in as additional fields in the "arrayData" or "hpaData" structures. This feature is temporary, but should be incorporated in a more elegant manner than its current form.